+++ packages/core/src/checkbox/styles/Wrapper.ts
@@ -1,11 +1,12 @@
+import { rem } from '@heathmont/moon-utils';
 import styled from 'styled-components';
 
-const Wrapper = styled.div(({ theme }) => ({
+const Wrapper = styled.div({
   display: 'flex',
   alignItems: 'center',
   justifyContent: 'center',
   position: 'relative',
-  gap: theme.newTokens.space.twoxsmall,
-}));
+  gap: rem(8),
+});
 
 export default Wrapper;