+++ packages/datepicker/src/DatePicker/private/month/styles/SelectorContainer.ts
@@ -9,7 +9,7 @@ const SelectorContainer = styled.div(({ theme: { newTokens, colorNew } }) => ({
   top: 0,
   right: 0,
   backgroundColor: colorNew.gohan,
-  borderRadius: newTokens.borderRadius.large,
+  borderRadius: newTokens.borderRadius.surface.medium,
   boxShadow: newTokens.boxShadow.large,
   padding: `${rem(6)} ${rem(8)}`,
 }));