+++ packages/themes/src/empireDark/empireDark.ts
@@ -77,13 +77,6 @@ const empireDark: Theme = {
     default: 8,
     largest: 8,
   },
-  newTokens: {
-    ...sharedTokens.newTokens,
-    borderRadius: {
-      ...sharedTokens.newTokens.borderRadius,
-      interactive: sharedTokens.newTokens.borderRadius.medium,
-    },
-  },
   ...componentsTokens(colorNew),
   colorScheme: 'dark',
 };