+++ packages/themes/src/livecasinoLight/livecasinoLight.ts
@@ -81,13 +81,6 @@ const livecasinoLight: Theme = {
     ...colorNew,
     ...supportiveColors,
   },
-  newTokens: {
-    ...sharedTokens.newTokens,
-    borderRadius: {
-      ...sharedTokens.newTokens.borderRadius,
-      interactive: sharedTokens.newTokens.borderRadius.medium,
-    },
-  },
   ...componentsTokens(colorNew),
   colorScheme: 'light',
 };