+++ workspaces/core/src/button/private/types/ButtonIcons.ts
@@ -1,0 +1,3 @@
+type ButtonIcons = 'left' | 'right' | 'only';
+
+export default ButtonIcons;